requirements.txt
selenium>=4.14.0
pytest>=9.0.0
allure-pytest>=2.0.0
pytest-html>=5.0.0
webdriver-manager>=3.8.0
